api/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
psycopg2-binary==2.9.3
pydantic==1.10.2
python-dotenv==0.21.0 